Update some code

+++ src/features/favorite/favorite.service.js
@@ -3,18 +3,30 @@ const ProductRepo = require('../product/product.repo');
 const { BadRequest } = require('../../utils/createError');
 
 module.exports = {
-    getWishList: async (userId) => {
-        return await FavoriteRepo.getWishList(userId);
+    getFavoriteOfUser: async (userId) => {
+        let favorite = await FavoriteRepo.getFavoriteFromCache(userId);
+        if(favorite) return favorite;
+        favorite = await FavoriteRepo.getFavoriteFromDb(userId);
+        await FavoriteRepo.writeFavoriteToCache(userId, favorite);
+        return 
     },
     addProductToList: async (userId, productId) => {
         const product = await ProductRepo.getProductById(productId);
         if(!product) throw BadRequest("Product is not existed");
-        const updatedFavorite = await FavoriteRepo.addProductToList(userId, product);
+        const favorite = await FavoriteRepo.getFavoriteFromDb(userId);
+        const existingItemIndex = favorite.items.findIndex(item => item.equals(product.id));
+        if (existingItemIndex !== -1) return await favorite.populate('items');
+        const updatedFavorite = await FavoriteRepo.addProductToList(favorite, product);
+        await FavoriteRepo.writeFavoriteToCache(userId, updatedFavorite);
         return updatedFavorite;
     },
     removeFromFavorite: async (userId, productId) => {
         const product = await ProductRepo.getProductById(productId);
         if (!product) throw BadRequest("Product is not existed");
-        return await FavoriteRepo.removeFromList(userId, productId);
+        const existingItemIndex = favorite.items.findIndex(item => item.equals(product.id));
+        if (existingItemIndex === -1) return await favorite.populate('items');
+        const updatedFavorite =  await FavoriteRepo.removeFromList(userId, productId);
+        await FavoriteRepo.writeFavoriteToCache(userId, updatedFavorite);
+        return updatedFavorite;
     }
 }
